+++ src/AtomUI.Desktop.Controls/Statistic/Statistic.SemanticParts.cs
@@ -1,0 +1,39 @@
+using AtomUI.Theme;
+using Avalonia.Controls;
+using Avalonia.Controls.Presenters;
+
+namespace AtomUI.Desktop.Controls;
+
+[SemanticPart(
+    "header",
+    SelectorClass = "semantic-header",
+    ContractType = typeof(Border),
+    Since = "6.0")]
+[SemanticPart(
+    "title",
+    SelectorClass = "semantic-title",
+    ContractType = typeof(ContentPresenter),
+    Since = "6.0")]
+[SemanticPart(
+    "content",
+    SelectorClass = "semantic-content",
+    ContractType = typeof(StackPanel),
+    Since = "6.0")]
+[SemanticPart(
+    "value",
+    SelectorClass = "semantic-value",
+    ContractType = typeof(ContentPresenter),
+    Since = "6.0")]
+[SemanticPart(
+    "prefix",
+    SelectorClass = "semantic-prefix",
+    ContractType = typeof(ContentPresenter),
+    Since = "6.0")]
+[SemanticPart(
+    "suffix",
+    SelectorClass = "semantic-suffix",
+    ContractType = typeof(ContentPresenter),
+    Since = "6.0")]
+public partial class Statistic
+{
+}